requirements.txt
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
